Rework sleep/wake messaging
Before, the burst message we receive while sleeping was unencrypted, so we
only used it as a mechanism to wake up and receive the "real" message. RFM69_WL
now supports encrypted burst messages, so we don't have to do this. Now we
can send the actual message in the burst, handle it as usual on the other
side and send an ACK command back.

There are also some miscellaneous cleanups. Initializing the radio before the flash seems to clear up
some problems with flash init hanging. We also stop using high-speed listen mode, as the range
was not really good enough in some cases. As a result, the RX time was raised to 512us.